Clean up peptide analysis module.
