+++ modules/PG_PeptideAnalysis/peptide_analysis.sh
@@ -1,1 +1,1 @@
-python ./peptide_analysis_v2.py -gmap ../../results/PG_ReferenceTables/gene_to_isoname.tsv -gc ../../data/AllPeptides_Gencode.psmtsv -pb ../../data/jurkat_orf_refined.fasta -sft ../../data/pacbio_6frm_database_gene_grouped.fasta
+python ./peptide_analysis.py -gmap ../../results/PG_ReferenceTables/gene_to_isoname.tsv -gc ../../data/AllPeptides_Gencode.psmtsv -pb ../../data/jurkat_orf_refined.fasta -sft ../../data/pacbio_6frm_database_gene_grouped.fasta
